all up and running

+++ lib/id-counter/id-counter-api/src/main/java/bs/lib/id/counter/api/service/CounterInternalService.java
@@ -6,5 +6,5 @@ import bs.lib.id.counter.model.entity.Counter;
 import java.util.List;
 
 public interface CounterInternalService {
-    Long getNextValue(Counter counter, List<RecordAttribute> attributes);
+    Integer getNextValue(Counter counter, List<RecordAttribute> attributes);
 }
