+++ lib/id-counter/id-counter-repository/src/main/java/bs/lib/id/counter/repository/jpa/CounterInstanceAttributeJPARepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import bs.lib.id.counter.model.entity.CounterInstanceAttribute;
 
 @Repository
-public interface CounterInstanceAttributeJPARepository extends JpaRepository<CounterInstanceAttribute,Long> {
+public interface CounterInstanceAttributeJPARepository extends JpaRepository<CounterInstanceAttribute,Integer> {
 }
