+++ lib/id-counter/id-counter-repository/src/main/java/bs/lib/id/counter/repository/jpa/CounterInstanceJPARepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import bs.lib.id.counter.model.entity.CounterInstance;
 
 @Repository
-public interface CounterInstanceJPARepository extends JpaRepository<CounterInstance, Long> {
+public interface CounterInstanceJPARepository extends JpaRepository<CounterInstance, Integer> {
 
 }
